api_server/mq_task/requirements.txt
celery==5.0.5
pymongo==3.11.2
chainer==7.7.0
Pillow==8.1.0